(PC-27682)[PRO] refactor: merge constants for default values
