+++ pro/src/components/Address/index.ts
@@ -1,4 +1,3 @@
 export { default as validationSchema } from './validationSchema'
 export { default as AddressSelect } from './Address'
-export { DEFAULT_ADDRESS_FORM_VALUES } from './constants'
 export type { Address } from './types'
