+++ pro/src/components/VenueForm/Accessibility/index.ts
@@ -1,3 +1,2 @@
 export { default as validationSchema } from './validationSchema'
 export { default as Accessibility } from './Accessibility'
-export { DEFAULT_ACCESSIBILITY_FORM_VALUES } from './constants'
