--- pro/src/components/VenueForm/Activity/constants.ts
@@ -1,5 +1,0 @@
-export const DEFAULT_ACTIVITY_FORM_VALUES = {
-  description: '',
-  venueType: '',
-  venueLabel: '',
-}
