+++ pro/src/components/VenueForm/Activity/index.ts
@@ -1,2 +1,1 @@
 export { default as validationSchema } from './validationSchema'
-export { DEFAULT_ACTIVITY_FORM_VALUES } from './constants'
