--- pro/src/components/VenueForm/Contact/constants.ts
@@ -1,6 +1,0 @@
-export const DEFAULT_CONTACT_FORM_VALUES = {
-  phoneNumber: '',
-  email: '',
-  webSite: '',
-  bookingEmail: '',
-}
