+++ pro/src/components/VenueForm/Contact/index.ts
@@ -1,2 +1,1 @@
 export { default as validationSchema } from './validationSchema'
-export { DEFAULT_CONTACT_FORM_VALUES } from './constants'
