--- pro/src/components/VenueForm/ImageUploaderVenue/constants.ts
@@ -1,6 +1,0 @@
-export const DEFAULT_IMAGE_UPLOADER_FORM_VALUES = {
-  isPermanent: false,
-  bannerMeta: undefined,
-  id: undefined,
-  bannerUrl: '',
-}
