+++ pro/src/components/VenueForm/ImageUploaderVenue/index.ts
@@ -1,2 +1,1 @@
 export { default as ImageUploaderVenue } from './ImageUploaderVenue'
-export { DEFAULT_IMAGE_UPLOADER_FORM_VALUES } from './constants'
