--- pro/src/components/VenueForm/Informations/constants.ts
@@ -1,8 +1,0 @@
-export const DEFAULT_INFORMATIONS_FORM_VALUES = {
-  comment: '',
-  isVenueVirtual: false,
-  name: '',
-  publicName: '',
-  siret: '',
-  departmentCode: '',
-}
