+++ pro/src/components/VenueForm/Informations/index.ts
@@ -1,2 +1,1 @@
 export { default as validationSchema } from './validationSchema'
-export { DEFAULT_INFORMATIONS_FORM_VALUES } from './constants'
