--- pro/src/components/VenueForm/ReimbursementPoint/constants.ts
@@ -1,3 +1,0 @@
-export const DEFAULT_REIMBURSEMENT_POINT_FORM_VALUES = {
-  reimbursementPointId: null,
-}
