--- pro/src/components/VenueForm/ReimbursementPoint/index.ts
@@ -1,1 +1,0 @@
-export { DEFAULT_REIMBURSEMENT_POINT_FORM_VALUES } from './constants'
