--- pro/src/components/VenueForm/WithdrawalDetails/constants.ts
@@ -1,4 +1,0 @@
-export const DEFAULT_WITHDRAWAL_FORM_VALUES = {
-  withdrawalDetails: '',
-  isWithdrawalAppliedOnAllOffers: false,
-}
