--- pro/src/pages/Offerers/Offerer/VenueV1/fields/PricingPoint/constant.ts
@@ -1,3 +1,0 @@
-export const DEFAULT_PRICING_POINT_FORM_VALUES = {
-  venueSiret: null,
-}
