+++ pro/src/screens/SignupJourneyForm/Offerer/constants.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_ADDRESS_FORM_VALUES } from 'components/Address'
+import { DEFAULT_ADDRESS_FORM_VALUES } from 'components/Address/constants'
 
 export const DEFAULT_OFFERER_FORM_VALUES = {
   siret: '',
